+++ scripts/cleanup.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+set -e
+rm -rf dist build *.egg-info
