+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pytest-frozen-uuids
-version = 0.3.1
+version = 0.3.3
 description = Deterministically frozen UUID's for your tests
 long_description = file:README.md
 long_description_content_type = text/markdown
@@ -14,40 +14,29 @@ classifiers =
     Programming Language :: Python :: 3
 
 [options]
+python_requires = >=3.6
+packages = find_namespace:
 package_dir =
     = .
-packages = find_namespace:
 install_requires =
     pytest>=3.0
 
 [options.packages.find]
 where = .
-exclude = tests
+include = pytest_frozen_uuids
 
 [options.entry_points]
 pytest11 =
-    frozen_uuids = pytest_frozen_uuids
+    frozen_uuids = pytest_frozen_uuids.__main__
 
 [flake8]
 max-line-length = 100
 extended-ignore = W504
-exclude =
-    .git,
-    .venv,
-    .eggs,
-    .tox,
-    __pycache__
+exclude = .git,.venv,.eggs,.tox,__pycache__
 
 [coverage:run]
 branch = True
-source = .
-omit =
-    .venv,
-    .eggs,
-    .tox,
-    __pycache__,
-    tests/*,
-    setup.py,
+source = pytest_frozen_uuids
 
 [tool:pytest]
 testpaths = tests
@@ -57,7 +46,7 @@ addopts = -v
 [mypy]
 ignore_missing_imports = True
 show_error_context = True
-files = pytest_frozen_uuids.py,src
+files = pytest_frozen_uuids
 
 [tool:isort]
 profile = black
